add new catagory backend and move the previous implementation into node directory
