+++ ipam/centralip/backend/example-centalip.conf
@@ -1,0 +1,20 @@
+{
+   "name":"mynet",
+   "cniVersion":"0.3.1",
+   "type":"ovs",
+   "ovsBridge":"br0",
+   "vtepIPs":[
+      "10.0.0.3"
+   ],
+   "isDefaultGateway": true,
+   "ipMasq": true,
+   "ipam":{
+       "type":"centralip",
+       "ip": "node"
+       "network":"10.245.0.0/16",
+       "subnetLen": 24,
+       "subnetMin": "10.245.5.0",
+       "subnetMax": "10.245.50.0",
+       "etcdURL": "127.0.0.1:2379"
+   }
+}
