+++ ipam/centralip/backend/utils/utils.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package centralip
+package utils
 
 import (
 	"encoding/binary"
